Place server and client code in separate directories within src
