+++ .env
@@ -1,1 +1,1 @@
-NODE_PATH=src/
+NODE_PATH=src/client
