+++ src/client/actions/nickname.js
