+++ src/client/components/FontSize.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 const FontSize = {
-  Normal: styled.div`font-size: normal;`,
+  Normal: styled.div`font-size: 1em;`,
   Larger: styled.div`font-size: larger;`
 };
