+++ src/client/components/Panel.js
