+++ src/client/components/form/FormError.js
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
+import styled from 'styled-components';
 
 const FormError = styled.p`
   display: block;
