+++ src/client/middlewares/loginView.js
