+++ src/client/middlewares/socket/configure.js
