+++ src/client/reducers/loginView.js
