+++ src/server/configureStore.js
@@ -1,6 +1,6 @@
+import middlewares from 'middlewares';
 import reducer from 'reducer';
 import { applyMiddleware, createStore } from 'redux';
-import middlewares from './middlewares';
 
 const configureStore = () =>
   createStore(reducer, applyMiddleware(...middlewares));
